+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Include methods in `UnitLevelKeywordIndentation`.
 - Include implicit attribute constructor calls in `ForbiddenMethod`.
 - Always enforce the `Attribute` suffix in `AttributeName`.
+- Allow the `Attribute` suffix to be omitted for attribute references in `MixedNames`.
 - Improve name resolution for declarations within types.
 - Improve type resolution for array accesses into variants.
 - Improve parsing and type modeling around `AnsiString` types with specified code pages.